Refactor package structure and extract preconditions to separate submodule
